+++ packages/visualizations-react/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.31.2](https://github.com/opendatasoft/ods-dataviz-sdk/compare/@opendatasoft/visualizations-react@0.31.2-beta.0...@opendatasoft/visualizations-react@0.31.2) (2026-02-19)
+
+
+### Bug Fixes
+
+* improve time scale validation robustness ([faeb63f](https://github.com/opendatasoft/ods-dataviz-sdk/commit/faeb63f116934d337565273965d5faaeeb3cd334))
+
+
+
+
+
+## [0.31.2-beta.0](https://github.com/opendatasoft/ods-dataviz-sdk/compare/@opendatasoft/visualizations-react@0.30.0...@opendatasoft/visualizations-react@0.31.2-beta.0) (2026-02-10)
+
+
+### Bug Fixes
+
+* prevent "too many ticks" crash with automatic time unit fallback ([f5c7df9](https://github.com/opendatasoft/ods-dataviz-sdk/commit/f5c7df9828c415b01ad7cab3553c26900d7c2c47))
+
+
+
+
+
 ## [0.31.1](https://github.com/opendatasoft/ods-dataviz-sdk/compare/@opendatasoft/visualizations-react@0.31.0...@opendatasoft/visualizations-react@0.31.1) (2026-01-28)
 
 **Note:** Version bump only for package @opendatasoft/visualizations-react